% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.tiles.R
\name{add_tilenum}
\alias{add_tilenum}
\title{Add Tile Numbers to Spatial Features}
\usage{
add_tilenum(x, mercator = TRUE, zoom = 18L)
}
\arguments{
\item{x}{A spatial features object.}

\item{mercator}{A logical value indicating whether the coordinates are in Mercator projection. Default is TRUE.}

\item{zoom}{An integer indicating the zoom level for the tile numbers. Default is 18.}
}
\value{
A spatial features object with added tile numbers.
}
\description{
This function adds tile numbers to a spatial features object based on its centroid coordinates.
}
\examples{
nc <- sf::st_read(system.file('shape/nc.shp', package = 'sf'))
add_tilenum(nc)

}
\keyword{internal}
