% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.report.R
\name{addBreak}
\alias{addBreak}
\title{Add a break to a rdocx element}
\usage{
addBreak(doc, landscape_continuous = FALSE)
}
\arguments{
\item{doc}{a rdocx element}

\item{landscape_continuous}{a boolean. Must be TRUE in case of landscape disposition to be continued after the break.
Default to FALSE}
}
\value{
a rdocx element with a break at the end
}
\description{
Add a bread to a rdocx element
}
\examples{
doc <- read_docx()
doc <- addTextElement(doc, 'Titre 1 : description of the dataset', style = 'heading 1')
doc <- addBreak(doc)
doc <- addTextElement(doc, 'Titre 2 : comparaison of the dataset', style = 'heading 1')
doc <- setLandscape(doc, add_break = TRUE)
doc <- addBreak(doc, landscape_continuous = TRUE)
doc <- addTextElement(doc, 'Titre 3 : plot', style = 'heading 1')
doc <- setLandscape(doc, add_break = FALSE)
print(doc, target = 'output.docx')

}
