% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addClass}
\alias{addClass}
\title{Add a class to an object}
\usage{
addClass(x, what)
}
\arguments{
\item{x}{An object}

\item{what}{The name of the class to add}
}
\value{
The object with the new class
}
\description{
This function adds a new class to an object.
}
\examples{
x <- 1L
addClass(x, 'foo')
# Output:
# [1] 1
# attr(,'class')
# [1] 'numeric' 'foo'
class(x)
# Output:
# [1] 'numeric' 'foo'

}
\keyword{internal}
