% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.leaflet.R
\name{addCustomControl}
\alias{addCustomControl}
\title{Add a custom control to a Leaflet map.}
\usage{
addCustomControl(
  map,
  position = "bottomright",
  id = "hollow",
  title = "",
  classes = "",
  content = ""
)
}
\arguments{
\item{map}{A Leaflet map object.}

\item{position}{A character string specifying the position of the control on the map (default is 'bottomright').}

\item{id}{A character string providing an identifier for the custom control (default is 'hollow').}

\item{title}{A character string representing the title of the custom control.}

\item{classes}{A character string containing additional CSS classes to be applied to the custom control.}

\item{content}{A character string defining the content of the custom control.}
}
\value{
Updated Leaflet map with the custom control added.
}
\description{
This function adds a custom control to a Leaflet map allowing users to customize the control with specified parameters.
}
\examples{
\dontrun{
map <- leaflet() |>
  addTiles() |>
  setView(lng = 0L, lat = 0L, zoom = 2L)
addCustomControl(map, position = 'topright', id = 'custom', title = 'Custom Control', content = 'This is a custom control.')
}

}
