% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.leaflet.R
\name{addGoogleTiles}
\alias{addGoogleTiles}
\title{Adds Google Tiles Layer}
\usage{
addGoogleTiles(
  map,
  type = c("roadmap", "satellite", "terrain"),
  layerId = NULL,
  group = NULL,
  ...
)
}
\arguments{
\item{map}{The Map widget}

\item{type}{String. Type of Tiles to display}

\item{layerId}{String. An optional unique ID for the layer}

\item{group}{String. An optional group name for the layer}

\item{...}{Optional Parameters required by the Google API described at \url{https://msdn.microsoft.com/en-us/library/ff701716.aspx}}
}
\description{
Adds Google Tiles Layer
}
\examples{
\dontrun{
library(leaflet)
map <- leaflet() |>
  addTiles() |>
  setView(lng = -71.06, lat = 42.36, zoom = 12L)
addGoogleTiles(map)
}

}
\seealso{
Get a Google Maps API Key: \url{https://msdn.microsoft.com/en-us/library/ff428642.aspx}
}
