% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.report.R
\name{addNewLine}
\alias{addNewLine}
\title{Add new line to rdocx element}
\usage{
addNewLine(doc)
}
\arguments{
\item{doc}{a rdocx element}
}
\value{
a rdocx element with a new line at the end
}
\description{
Add a new line to rdocx element
}
