% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.leaflet.R
\name{addPaintPolygonControl}
\alias{addPaintPolygonControl}
\title{Add a PaintPolygon Control}
\usage{
addPaintPolygonControl(
  map,
  position = "bottomright",
  radius = 30L,
  minRadius = 10L,
  maxRadius = 50L,
  layerOptions = list(),
  drawOptions = list(weight = 1L),
  eraseOptions = list(color = "#ff324a", weight = 1L),
  menu = TRUE
)
}
\arguments{
\item{map}{The leaflet map widget.}

\item{position}{The position of the control (default is 'bottomright').}

\item{radius}{The radius of the brush (default is 30).}

\item{minRadius}{The minimum radius of the brush (default is 10).}

\item{maxRadius}{The maximum radius of the brush (default is 50).}

\item{layerOptions}{Additional options for the painted layer (default is an empty list).}

\item{drawOptions}{Options for drawing with the brush (default is a list with weight = 1).}

\item{eraseOptions}{Options for erasing with the brush (default is a list with color = '#ff324a' and weight = 1).}

\item{menu}{Logical, whether to display the menu options (default is TRUE).}
}
\description{
Add a PaintPolygon control to a leaflet map widget.
}
