% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.pkg.R
\name{addRoofsGoogle}
\alias{addRoofsGoogle}
\title{Add Roofs using Google Maps API}
\usage{
addRoofsGoogle(polygons, dir = getDirAppTemp(), async_queue = NULL)
}
\arguments{
\item{polygons}{A set of polygons in 'sfc_POLYGON' or 'sfc_MULTIPOLYGON' format.}

\item{dir}{The directory to save the retrieved images.}

\item{async_queue}{An optional asynchronous queue for progress tracking.}
}
\value{
A list containing the tiles and roofs retrieved from the Google Maps API.
}
\description{
This function takes a set of polygons and retrieves roof data using the Google Maps API.
}
\examples{
\dontrun{
  polygon <-  sf::st_set_crs(sf::st_sfc(st_bbox_polygon(c(13.105, 11.825, 13.11, 11.83))), 4326L)

  addRoofsGoogle(polygons = polygon)
}

}
