% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{angle}
\alias{angle}
\title{Calculate the angle between three points in a two-dimensional Cartesian coordinate system.}
\usage{
angle(x0, y0, x1, y1, x2, y2)
}
\arguments{
\item{x0}{The x-coordinate of the first point.}

\item{y0}{The y-coordinate of the first point.}

\item{x1}{The x-coordinate of the second point.}

\item{y1}{The y-coordinate of the second point.}

\item{x2}{The x-coordinate of the third point.}

\item{y2}{The y-coordinate of the third point.}
}
\value{
The angle in degrees between the three points.
}
\description{
This function calculates the angle in degrees between three points (x0, y0), (x1, y1), and (x2, y2) in a two-dimensional Cartesian coordinate system.
}
\examples{
angle(0L, 0L, 1L, 0L, 0L, 1L)
# Returns 45

angle(0L, 0L, 1L, 1L, 0L, 1L)
# Returns 90

}
\keyword{internal}
