% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ask}
\alias{ask}
\title{Prompt the user for input and return the entered value.}
\usage{
ask(prompt)
}
\arguments{
\item{prompt}{A character string specifying the prompt to display.}
}
\value{
A character string with the user input.
}
\description{
Prompt the user for input and return the entered value.
}
\examples{
ask('Enter your name: ')
ask('Enter a number: ')

}
\keyword{internal}
