% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_dropClass}
\alias{assert_dropClass}
\title{Assertive check for dropClass}
\usage{
assert_dropClass(x, what)
}
\arguments{
\item{x}{An object}

\item{what}{The class to be dropped}
}
\description{
This function checks if a given object belongs to a specific class and raises an error if it does.
}
\examples{
assert_dropClass('abc', 'character')
# Error: !any(class('abc') == 'character') is not TRUE
assert_dropClass(123L, 'numeric')
# Error: !any(class(123) == 'numeric') is not TRUE

}
\keyword{internal}
