% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.base.R
\name{Base}
\alias{Base}
\title{Base R6 Class}
\value{
An R6 class object.
}
\description{
This function defines an R6 class called "Base". The class
has a single public method called "initialize", which creates
a new instance of the class and sets its "parent" attribute to
the value passed in as an argument (or NULL if no argument is provided).

If a global option called "epi.log" is set to TRUE, all functions defined in the
class are decorated with logging functionality. The class also has a single
active field called "parent", which can be used to get or set the value of
the "parent" attribute.
}
\examples{
# Create a new instance of the Base class with no parent
my_base <- Base$new()

# Create a new instance of the Base class with a parent
my_child_base <- Base$new(parent = my_base)

# Get the value of the "parent" attribute for an instance of the Base class
my_base$parent()

# Set the value of the "parent" attribute for an instance of the Base class
my_base$parent(my_child_base)

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parent}}{(\code{R6Class})\cr
Parent object of the object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Base-new}{\code{Base$new()}}
\item \href{#method-Base-clone}{\code{Base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-new"></a>}}
\if{latex}{\out{\hypertarget{method-Base-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$new(parent = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{(\code{R6Class})\cr
Identifier of the parent object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Base-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Base-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
