% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{button}
\alias{button}
\title{Create a button widget}
\usage{
button(..., loading = FALSE, loading_label = .("Loading..."))
}
\arguments{
\item{...}{Additional arguments to be passed to the button function}

\item{loading}{Logical indicating if the button is in a loading state}

\item{loading_label}{The label to display when the button is in a loading state}

\item{inputId}{The input ID of the button}

\item{label}{The text to display on the button}

\item{icon}{An optional icon to display alongside the label}

\item{class}{An optional class for styling the button}
}
\value{
A button widget with specified characteristics
}
\description{
This function creates a button with specified characteristics.
}
\examples{
button('button_id', 'Click me', icon = 'check')
button('submit', 'Submit', loading = TRUE)

}
