% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{buttonLoading}
\alias{buttonLoading}
\title{Create a button with optional loading state and loading label}
\usage{
buttonLoading(
  ...,
  margin = 2L,
  semantic = "default",
  size = NULL,
  hide = FALSE,
  outline = FALSE
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[shinyFeedback:loadingButton]{shinyFeedback::loadingButton}}
  \describe{
    \item{\code{inputId}}{the input id}
    \item{\code{label}}{the button text (label)}
    \item{\code{class}}{the class(es) to apply to the button}
    \item{\code{style}}{style for button (pre-loading); character string w/ CSS styling format: "color: black; background-color: red;"}
    \item{\code{loadingLabel}}{text to show after button is clicked (e.g. during loading)}
    \item{\code{loadingSpinner}}{the loading spinner icon.  Valid values are NULL, "spinner", "circle-notch",
"sync", and "cog"}
    \item{\code{loadingClass}}{the loading button css class(es).}
    \item{\code{loadingStyle}}{style for button (while loading); character string w/ CSS styling format: "color: black; background-color: red;"}
  }}
}
\value{
HTML button with optional loading behavior
}
\description{
Create a button with optional loading state and loading label
}
\examples{
ui <- tagList(
  buttonLoading('act_submit', label = 'Submit'),
  buttonLoading('act_submit', label = 'Submit', semantic = 'danger')
)

shinyAppMinimal(ui = ui)

}
