% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c_named}
\alias{c_named}
\title{Assign Names to a Vector}
\usage{
c_named(x, names)
}
\arguments{
\item{x}{A vector.}

\item{names}{A character vector specifying the names to assign to the elements of the vector \code{x}.}
}
\value{
A vector with assigned names.
}
\description{
This function assigns names to elements of a vector.
}
\examples{
c_named(c(1L, 2L, 3L), c('one', 'two', 'three'))

}
\keyword{internal}
