% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.pkg.R
\name{calculateSampleZeroInflatedDistribution}
\alias{calculateSampleZeroInflatedDistribution}
\title{Calculate Sample Zero-Inflated Distribution}
\usage{
calculateSampleZeroInflatedDistribution(
  smp_pop,
  n = 10000L,
  error_confidence = 95L,
  progress = NULL
)
}
\arguments{
\item{smp_pop}{numeric vector of the sample data}

\item{n}{integer value indicating the number of bootstrap samples to take (default is 10000)}

\item{error_confidence}{integer value indicating the error confidence level (default is 95)}

\item{progress}{a progress bar object created by \code{\link[progress]{progress_bar}} (default is NULL)}
}
\value{
a list with the following elements:
\item{pop}{integer value of the estimated population}
\item{pop_i}{integer value of the population confidence interval}
\item{lambda}{numeric value of the estimated lambda parameter}
\item{prob}{numeric value of the estimated probability of zero inflation}
}
\description{
This function calculates the population estimate and confidence interval for a sample zero-inflated distribution.
}
\examples{
data <- c(1L, 2L, 2L, 3L, 3L, 3L, 4L, 4L, 5L, 5L, 5L, 5L, 6L, 6L, 7L, 7L, 7L, 7L, 7L, 7L, 8L, 8L, 8L, 8L, 9L, 9L, 9L, 9L, 10L, 10L)
calculateSampleZeroInflatedDistribution(data)

}
