% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catf}
\alias{catf}
\title{Logging}
\usage{
catf(fmt, ...)
}
\arguments{
\item{fmt}{A character string containing format specifications.}

\item{...}{Optional arguments to be formatted.}
}
\description{
This function prints formatted text to the console.
}
\examples{
catf('Hello, \%s', 'world!')
# Output: Hello, world!

}
