% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{checkPort}
\alias{checkPort}
\title{Function to check if a port is open on localhost}
\usage{
checkPort(port, timeout = 0.1)
}
\arguments{
\item{port}{Port number to be checked}

\item{timeout}{Timeout value for the connection attempt (default is 0.1 seconds)}
}
\value{
Logical value indicating if the port is open
}
\description{
Function to check if a port is open on localhost
}
\examples{
checkPort(80L)
checkPort(443L, timeout = 0.5)

}
