% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.paths.R
\name{cleanDirCache}
\alias{cleanDirCache}
\title{Clean directory cache}
\usage{
cleanDirCache()
}
\description{
This function deletes the cache folder in the current working directory.
}
\examples{
cleanDirCache()

}
\keyword{internal}
