% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2hex}
\alias{col2hex}
\title{Convert Color Names to Hexadecimal Format}
\usage{
col2hex(col, alpha = FALSE)
}
\arguments{
\item{col}{a character vector specifying the color name(s) to be converted.}

\item{alpha}{a logical value indicating whether to include alpha channel in the resulting hexadecimal color(s). Default is \code{FALSE}.}
}
\value{
a character vector of the color(s) in hexadecimal format.
}
\description{
This function takes a color name and converts it to hexadecimal format.
}
\examples{
col2hex('red') # '#FF0000'
col2hex(c('blue', 'green')) # c('#0000FF', '#00FF00')

}
\keyword{internal}
