% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse}
\alias{collapse}
\title{Collapse a character vector into a single string with elements separated by a delimiter.}
\usage{
collapse(x, quote = TRUE, sep = ", ")
}
\arguments{
\item{x}{A character vector to be collapsed.}

\item{quote}{If \code{TRUE}, each element of \code{x} will be quoted using double quotes. Default is \code{TRUE}.}

\item{sep}{The delimiter to be used to separate the elements of \code{x}. Default is \code{', '}.}
}
\value{
A single string with collapsed elements separated by the specified delimiter.
}
\description{
This function takes a character vector and collapses it into a single string by concatenating its elements with a specified delimiter. By default, each element is quoted using double quotes. The delimiter can be specified using the \code{sep} parameter.
}
\examples{
collapse(c('a', 'b', 'c'))
# Output: 'a, b, c'

collapse(c('a', 'b', 'c'), quote = FALSE, sep = '-')
# Output: a-b-c

}
\seealso{
\code{\link{paste}}
}
\keyword{collapse}
\keyword{internal}
\keyword{string}
