% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare}
\alias{compare}
\title{Compare two files using Beyond Compare}
\usage{
compare(source, target, remote = FALSE)
}
\arguments{
\item{source}{A character string representing the path to the source file.}

\item{target}{A character string representing the path to the target file.}
}
\description{
This function compares two files using Beyond Compare. It converts the source and target to strings and passes them as arguments to the system command 'bcompare' along with the necessary options for a comparison using SFTP protocol.
}
\examples{
compare('sourcefile', 'targetfile')

}
\keyword{internal}
