% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{console.out}
\alias{console.out}
\title{Console output function}
\usage{
console.out(msg, ...)
}
\arguments{
\item{msg}{The message to print. Can include formatting code.}

\item{...}{Additional arguments for formatting the message.}
}
\description{
This function prints messages to the console with a timestamp and optional formatting.
The function also records the time elapsed since the last call to \code{console.out}.
}
\keyword{internal}
