% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degrees}
\alias{degrees}
\title{Convert radians to degrees}
\usage{
degrees(angle)
}
\arguments{
\item{angle}{The angle in radians.}
}
\value{
The angle in degrees.
}
\description{
This function takes an angle in radians and converts it to degrees.
}
\examples{
degrees(pi / 2L)
# Output: 90

}
\keyword{internal}
