% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.paths.R
\name{deleteDirCache}
\alias{deleteDirCache}
\title{Delete the directory cache}
\usage{
deleteDirCache()
}
\description{
This internal function deletes the directory cache using the \code{fs::dir_delete}
function and the \code{getDirApp} function to get the directory path.
}
\keyword{internal}
