% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.paths.R
\name{download}
\alias{download}
\title{Download a file from a URL.}
\usage{
download(url, destfile = fs::path_file(url), destdir = fs::path_temp())
}
\arguments{
\item{url}{The URL of the file to be downloaded.}

\item{destfile}{The name of the file to save the downloaded content to. By default, it generates a file name based on the URL.}

\item{destdir}{The destination directory where the downloaded file will be saved. By default, it uses the temporary directory.}
}
\value{
The path of the downloaded file.
}
\description{
This function downloads a file from a given URL and saves it to a specified destination directory.
}
\examples{
# Download a file and save it to the current working directory
download('http://example.com/file.txt')

# Download a file and save it to a specific directory with a custom file name
download('http://example.com/file.txt', 'data.txt', '/path/to/directory')
}
\keyword{internal}
