% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropClass}
\alias{dropClass}
\title{Drop a class from an object}
\usage{
dropClass(x, what)
}
\arguments{
\item{x}{An object}

\item{what}{The class to drop}
}
\value{
An object with the specified class removed
}
\description{
This function can be used to drop a specific class from an object.
}
\examples{
x <- structure(1L, class = c('a', 'b'))
dropClass(x, 'b')

}
\keyword{internal}
