% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropNA}
\alias{dropNA}
\title{Drops NA Values}
\usage{
dropNA(x)
}
\arguments{
\item{x}{The input vector}
}
\value{
A vector with NA values removed
}
\description{
This function takes a vector and removes any NA values from it.
}
\examples{
dropNA(c(1L, 2L, NA, 3L, NA, 4L))
# [1] 1 2 3 4

dropNA(c('a', 'b', NA, 'c', NA, 'd'))
# [1] 'a' 'b' 'c' 'd'

dropNA(c(TRUE, FALSE, NA, TRUE, NA, FALSE))
# [1]  TRUE FALSE  TRUE FALSE

}
\keyword{internal}
