% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropNulls}
\alias{dropNulls}
\title{Drop Null Values}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{The input vector or list.}
}
\value{
A vector or list with null values removed.
}
\description{
This function removes null values from a given input.
}
\examples{
dropNulls(c(1L, NULL, 3L, NULL))
# Output: [1] 1 3

dropNulls(list('a', NULL, 'b', NULL))
# Output: [[1]] [1] 'a'
#         [[2]] [1] 'b'

}
\keyword{internal}
