% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropNullsOrEmpty}
\alias{dropNullsOrEmpty}
\title{Drop Nulls or Empty Values}
\usage{
dropNullsOrEmpty(x)
}
\arguments{
\item{x}{A vector or a list.}
}
\value{
A vector or a list with the NULL or empty values removed.
}
\description{
Remove elements that are either NULL or empty from a vector or list.
}
\examples{
dropNullsOrEmpty(c(1L, 2L, NULL, '', 5L))
# [1] 1 2 5

dropNullsOrEmpty(list(1L, 2L, NULL, '', 5L))
# [[1]]
# [1] 1
#
# [[2]]
# [1] 2
#
# [[3]]
# [1] 5

}
\keyword{internal}
