% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explore}
\alias{explore}
\title{Open File Explorer at Specified Path}
\usage{
explore(path = fs::path_home())
}
\arguments{
\item{path}{A character string specifying the path to open File Explorer at. Default is the current working directory.}
}
\value{
This function is used for its side effects and does not return any value.
}
\description{
This function opens the File Explorer at the specified path.
}
\examples{
explore()
explore('C:/Users/')

}
\keyword{internal}
