% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.paths.R
\name{exploreDirCache}
\alias{exploreDirCache}
\title{Explore the directory cache}
\usage{
exploreDirCache()
}
\description{
This function explores the directory cache by executing the \code{explore()} function on the current application directory.
}
\keyword{internal}
