% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getChoices}
\alias{getChoices}
\title{Get choices for labelling}
\usage{
getChoices(df_values, var_name = "label", var_code = NULL, overall = NULL)
}
\arguments{
\item{df_values}{A data frame of values.}

\item{var_name}{The variable name for the labels. Default is 'label'.}

\item{var_code}{The variable code for the choices. If not provided, row names are used.}

\item{overall}{The overall label for all choices. If provided, it is added to the list of choices with a code of 'T'.}
}
\value{
A list of choices with corresponding labels.
}
\description{
This function is used to generate choices for labelling purposes. It takes in a data frame of values and outputs a list of choices with corresponding labels.
}
\examples{
df <- data.frame(var1 = c('A', 'B', 'C'), var2 = c(1L, 2L, 3L))
choices <- getChoices(df, 'var1', 'var2', overall = 'Overall')
print(choices)
}
\seealso{
\code{\link{label}} for labelling.
}
