% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCodeFactors}
\alias{getCodeFactors}
\title{Get coded factors from codes and values.}
\usage{
getCodeFactors(codes, values, short = FALSE)
}
\arguments{
\item{codes}{A data frame or tibble containing code and label information.}

\item{values}{A vector of values to be converted into factors.}

\item{short}{Logical value indicating whether to use short labels (default is \code{FALSE}).}
}
\value{
A factor variable created from the values and codes.
}
\description{
This function takes two input vectors, \code{codes} and \code{values}, and returns a factor variable based on the values and codes. The function can also take an additional argument \code{short} which, when set to \code{TRUE}, returns factors with short labels. By default, the function returns factors with full labels.
}
\examples{
codes <- data.frame(cde_code = c(1L, 2L, 3L), cde_label = c('Low', 'Medium', 'High'), cde_label_short = c('L', 'M', 'H'))
values <- c(1L, 3L)
getCodeFactors(codes, values)
getCodeFactors(codes, values, short = TRUE)
}
\keyword{internal}
