% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCodePath}
\alias{getCodePath}
\title{Get code path}
\usage{
getCodePath(codes, values)
}
\arguments{
\item{codes}{A data frame containing code information.}

\item{values}{A vector of values for which code paths are needed.}
}
\value{
A vector of code paths for each value in \code{values}.
}
\description{
This function takes in two arguments, \code{codes} and \code{values}, and returns the code path for the given values.
}
\details{
This function replaces all \code{NA} values in \code{paths} with '.' and then retrieves the code path for each value in \code{values} using the matching \code{cde_code} in \code{codes}.
}
\examples{
codes <- data.frame(cde_code = c(1L, 2L, 3L), cde_path = c('path1', 'path2', 'path3'))
values <- c(2L, 1L)
getCodePath(codes, values)
}
\keyword{internal}
