% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.data.R
\name{getCodeValues}
\alias{getCodeValues}
\title{Get code values}
\usage{
getCodeValues(codes, labels, short = FALSE)
}
\arguments{
\item{codes}{A data frame containing the codes and labels.}

\item{labels}{A vector of labels for which the code values should be retrieved.}

\item{short}{A logical value indicating whether the short labels should be used. Default is \code{FALSE}.}
}
\value{
A vector of corresponding code values for the given labels.
}
\description{
This function retrieves the code values for a given list of labels from a codes data frame.
}
\examples{
codes <- data.frame(
  cde_code = c(1L, 2L, 3L), cde_label = c('Label 1', 'Label 2', 'Label 3'),
  cde_label_short = c('L1', 'L2', 'L3')
)
labels <- c('Label 1', 'Label 2', NA, 'Label 3')
getCodeValues(codes, labels) # returns: 1, 2, NA, 3
getCodeValues(codes, labels, short = TRUE) # returns: L1, L2, NA, L3
}
\keyword{internal}
