% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFileCache}
\alias{getFileCache}
\title{Get File Cache}
\usage{
getFileCache(name, ext = "feather")
}
\arguments{
\item{name}{The name of the file cache.}

\item{ext}{The extension of the file cache. Default value is 'feather'.}
}
\value{
The path of the specified file cache.
}
\description{
This function returns the path of the specified file cache.
}
\keyword{internal}
