% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.string.R
\name{getFileExt}
\alias{getFileExt}
\title{Return the file extension.}
\usage{
getFileExt(path, tolower = TRUE)
}
\arguments{
\item{path}{The file path for which the extension needs to be determined.}

\item{tolower}{Logical, indicating whether the returned extension should be converted to lowercase. Default is \code{TRUE}.}
}
\value{
The file extension as a character string.
}
\description{
This function takes a file path as input and returns the file extension. The file extension is defined as the characters after the last period (.) in the file name.
}
\examples{
getFileExt('path/to/file.txt') # Returns 'txt'
getFileExt('path/to/file.TXT', tolower = FALSE) # Returns 'TXT'

}
\keyword{internal}
