% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFilters}
\alias{getFilters}
\title{Get Filters}
\usage{
getFilters()
}
\value{
A list of filters used for data filtering.
}
\description{
A function to retrieve the filters used for data filtering.
}
\keyword{internal}
