% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.data.R
\name{getGroupedChoices}
\alias{getGroupedChoices}
\title{Get Grouped Choices}
\usage{
getGroupedChoices(
  df_values,
  var_name = "label",
  var_code = NULL,
  var_group = NULL,
  overall = NULL
)
}
\arguments{
\item{df_values}{A dataframe containing values, names, and groups.}

\item{var_name}{The name of the column containing choice names.}

\item{var_code}{The name of the column containing choice values. If NULL, row names are used.}

\item{var_group}{The name of the column containing choice groups.}

\item{overall}{The name of the overall group. If specified, an overall group will be included in the output.}
}
\value{
A named list of choices, grouped by their respective groups.
}
\description{
This function takes in a dataframe containing values, names, and groups,
and returns choices grouped by their respective groups. It also allows for an overall
group to be specified and included in the output.
}
\examples{
df <- data.frame(
  var_code = c(1L, 2L, 3L, 4L),
  var_name = c('Choice A', 'Choice B', 'Choice C', 'Choice D'),
  var_group = c('Group 1', 'Group 1', 'Group 2', 'Group 2')
)
getGroupedChoices(df)

}
\keyword{internal}
