% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLabels}
\alias{getLabels}
\title{Get Labels}
\usage{
getLabels(df_values, code, var_name = "label")
}
\arguments{
\item{df_values}{A data frame containing the values and labels.}

\item{code}{The code used to identify the label.}

\item{var_name}{The name of the variable which contains the labels. Default is 'label'.}
}
\value{
A vector of labels corresponding to the given code.
}
\description{
This function extracts labels from a data frame based on a code and variable name.
}
\examples{
df <- data.frame(code = c('A', 'B', 'C'), label = c('Label A', 'Label B', 'Label C'))
getLabels(df, 'B')

}
\keyword{internal}
