% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPackageField}
\alias{getPackageField}
\title{Get a specific field from the package description}
\usage{
getPackageField(field = NULL)
}
\arguments{
\item{field}{The field to retrieve from the package description. If NULL, the entire package description is returned.}
}
\value{
The value of the specified field. If the field is not found, NULL is returned.
}
\description{
This function retrieves a specific field from the package description file.
}
\examples{
getPackageField('Title')
getPackageField('Version')
getPackageField()

}
\keyword{internal}
