% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPackageIdentifier}
\alias{getPackageIdentifier}
\title{Get Package Identifier}
\usage{
getPackageIdentifier()
}
\value{
The package identifier as a character string.
}
\description{
This function returns the identifier for a package by converting the package name to lowercase and replacing any '.' or '-' with '_'.
}
\details{
The function uses the packageName function from the utils package to get the current package name.
It then uses the str_replace_all and str_to_lower functions from the stringr package to convert the package name to lowercase and replace any '.' or '-' with '_'.
The resulting identifier is then returned.
}
\examples{
getPackageIdentifier()

}
\keyword{internal}
