% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPackagePath}
\alias{getPackagePath}
\title{Wrapper to fs::path_package for the package}
\usage{
getPackagePath(...)
}
\arguments{
\item{...}{Other parameters to be passed to \link[fs:path_package]{fs::path_package}.}
}
\value{
The path to the installed package.
}
\description{
This function is a wrapper for the \code{fs::path_package} function. It returns the path to the installed package.
}
\details{
This function internally calls \link[fs:path_package]{fs::path_package} and returns the package path by providing the current package name using \code{utils::packageName()}.
}
\examples{
getPackagePath('data.csv')

}
\seealso{
\link[fs:path_package]{fs::path_package} which this function wraps.
}
\keyword{internal}
