% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPackageTitle}
\alias{getPackageTitle}
\title{Get the package title}
\usage{
getPackageTitle()
}
\value{
The package title as a character string.
}
\description{
This function retrieves the value of the 'Title' field in a package description file.
}
\keyword{internal}
