% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPackageTitleL}
\alias{getPackageTitleL}
\title{Get Package Title in Lowercase}
\usage{
getPackageTitleL()
}
\value{
The package title in lowercase.
}
\description{
This function returns the title of a package in lowercase.
}
\details{
This function requires the following packages to be installed: \code{tools}.
}
\examples{
getPackageTitleL()
## 'my package'

}
\keyword{internal}
