% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPackageVersion}
\alias{getPackageVersion}
\title{Get Package Version}
\usage{
getPackageVersion(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[utils:packageDescription]{utils::packageVersion}}
  \describe{
    \item{\code{lib.loc}}{a character vector of directory names of \R libraries,
    or \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.  If the default is used, the loaded
    packages and namespaces are searched before the libraries.}
  }}
}
\value{
A character string representing the version of the package.
}
\description{
This function returns the version of the package.
}
\examples{
getPackageVersion()

}
\keyword{internal}
