% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getShinyServerHost}
\alias{getShinyServerHost}
\title{Get the Shiny Server host}
\usage{
getShinyServerHost(session)
}
\arguments{
\item{session}{The Shiny session object.}
}
\value{
The hostname (character) from the client's URL.
}
\description{
Retrieves the hostname from the Shiny session's client data.
}
\keyword{internal}
