% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{getTablePorts}
\alias{getTablePorts}
\title{Get table ports}
\usage{
getTablePorts()
}
\value{
A tibble containing local and remote port information
}
\description{
Get table ports
}
\examples{
getTablePorts()

}
