% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%empty\%}
\alias{\%empty\%}
\title{Empty Operator}
\usage{
a \%empty\% b
}
\arguments{
\item{a}{The first input.}

\item{b}{The second input.}
}
\value{
The non-empty input between \code{a} and \code{b}.
}
\description{
This function returns the non-empty input among two given inputs.
}
\examples{
x <- 'Hello'
y <- ''
z <- x \%empty\% y
# z will be 'Hello'

x <- ''
y <- 'World'
z <- x \%empty\% y
# z will be 'World'

x <- ''
y <- ''
z <- x \%empty\% y
# z will be ''

}
\keyword{internal}
