% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info}
\alias{info}
\title{bash alias for \code{\link[utils]{sessionInfo}}}
\usage{
info(package = NULL)
}
\arguments{
\item{package}{a character vector naming installed packages, or \code{NULL}
   (the default) meaning all attached packages.}
}
\description{
bash alias for \code{\link[utils]{sessionInfo}}
}
