% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.all.whole.number}
\alias{is.all.whole.number}
\title{Check if all elements in a vector are whole numbers}
\usage{
is.all.whole.number(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
TRUE if all elements in \code{x} are whole numbers, FALSE otherwise
}
\description{
This function takes a numeric vector and returns TRUE if all elements in the vector are whole numbers, and FALSE otherwise.
}
\examples{
is.all.whole.number(c(1L, 2L, 3L)) # should return TRUE
is.all.whole.number(c(1.5, 2.5, 3.5)) # should return FALSE

}
\keyword{internal}
