% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.defined}
\alias{is.defined}
\title{Check if an object is defined}
\usage{
is.defined(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
A logical value indicating if the object is defined or not.
}
\description{
This function checks if an object is defined in the R environment.
}
\examples{
is.defined(a)
# FALSE
a <- 5L
is.defined(a)
# TRUE

}
\keyword{internal}
