% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.dev}
\alias{is.dev}
\title{Check if the package is in development mode}
\usage{
is.dev()
}
\value{
A logical value indicating whether the package is in development mode or not
}
\description{
This function checks if the current package is in development mode
}
\examples{
is.dev()

}
\keyword{internal}
