% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.error}
\alias{is.error}
\title{Determine if an object is a try-error}
\usage{
is.error(x)
}
\arguments{
\item{x}{An object to check for being a try-error}
}
\value{
\code{TRUE} if \code{x} is a try-error, \code{FALSE} otherwise
}
\description{
This function takes an object as input and returns \code{TRUE} if the object
is a try-error object and \code{FALSE} otherwise.
}
\examples{
is.error(tryCatch(stop('An error occurred'), error = function(e) e))
# Returns TRUE

is.error(10L)
# Returns FALSE

is.error(NULL)
# Returns FALSE

}
\keyword{internal}
