% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.not.null}
\alias{is.not.null}
\title{Check if an object is not null}
\usage{
is.not.null(x)
}
\arguments{
\item{x}{The object to be checked}
}
\value{
A logical value indicating whether the object is not null
}
\description{
This function checks if an R object is not null.
}
\examples{
is.not.null(NULL)
# FALSE

is.not.null(1L)
# TRUE

}
\keyword{internal}
