% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.null.or.empty}
\alias{is.null.or.empty}
\title{Checks if a given object is NULL, empty, or containing only empty strings.}
\usage{
is.null.or.empty(x)
}
\arguments{
\item{x}{The object to be checked.}
}
\value{
A logical value indicating if the object is NULL, empty, or containing only empty strings.
}
\description{
This function returns TRUE if the object is NULL, has a length of 0, or is an empty string ('').
Otherwise, it returns FALSE.
}
\examples{
is.null.or.empty(NULL)
# TRUE
is.null.or.empty(character())
# TRUE
is.null.or.empty('')
# TRUE
is.null.or.empty(0L)
# FALSE
}
\keyword{internal}
