% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.numeric.character}
\alias{is.numeric.character}
\title{Check if values in a vector are numeric or character}
\usage{
\method{is.numeric}{character}(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
A logical vector indicating whether each element in \code{x} is numeric or character
}
\description{
This function determines whether the values in a vector are numeric or character. It can handle atomic vectors or lists.
}
\keyword{internal}
