% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.unique}
\alias{is.unique}
\title{Check if a vector has only unique elements}
\usage{
is.unique(x)
}
\arguments{
\item{x}{A vector of any type}
}
\value{
TRUE if all elements of the vector are unique, FALSE otherwise.
}
\description{
This function takes a vector as input and checks if it contains only unique elements.
}
\examples{
is.unique(c(1L, 2L, 3L)) # TRUE
is.unique(c(1L, 1L, 2L)) # FALSE

}
\keyword{internal}
