% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.win}
\alias{is.win}
\title{Check if the current operating system is Windows}
\usage{
is.win()
}
\value{
\code{TRUE} if the current operating system is Windows, \code{FALSE} otherwise.
}
\description{
This function checks if the current operating system is Windows.
}
\examples{
is.win()

}
\keyword{internal}
