% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.x32}
\alias{is.x32}
\title{Check if the R platform is 32-bit.}
\usage{
is.x32()
}
\value{
\code{TRUE} if the R platform is 32-bit, \code{FALSE} otherwise.
}
\description{
This function checks if the R platform is 32-bit by comparing the value of the 'r_arch' component of the .Platform list to 'x32'.
}
\examples{
is.x32()

}
\keyword{internal}
