% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.x64}
\alias{is.x64}
\title{Check if the R platform is x64}
\usage{
is.x64()
}
\value{
\code{TRUE} if the R platform is x64, \code{FALSE} otherwise.
}
\description{
This function checks if the R platform is x64.
}
\examples{
is.x64() # Returns TRUE if R platform is x64

}
\keyword{internal}
