% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isDevPackage}
\alias{isDevPackage}
\title{Function to check if R package is in development mode}
\usage{
isDevPackage()
}
\value{
a logical value indicating if the package is in development mode
}
\description{
This function checks if the current R package is in development mode by
examining the package description file. It looks for a specific pattern
in the version number, which indicates if it is a development version.
}
\examples{
isDevPackage()

}
\keyword{internal}
