% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isHostedApp}
\alias{isHostedApp}
\title{Check if the app is hosted on Shiny Server}
\usage{
isHostedApp()
}
\value{
Returns TRUE if the app is hosted on Shiny Server and FALSE otherwise.
}
\description{
This function checks whether the Shiny Server environment variable 'SHINY_SERVER_VERSION'
is set and returns TRUE only if it is true, and the application is running in a shiny session.
}
\examples{
isHostedApp()
# expected output: FALSE

}
\keyword{internal}
