% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isLoadedInDev}
\alias{isLoadedInDev}
\title{Check if a package is loaded in the development environment}
\usage{
isLoadedInDev()
}
\value{
A logical value indicating if the package is loaded in the development environment.
}
\description{
This function checks if a package called 'devtools_shims' is loaded in the development environment by searching for its associated file using the \code{utils::find} function.
}
\examples{
isLoadedInDev()
# [1] TRUE

}
\keyword{internal}
