% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isShinyApp}
\alias{isShinyApp}
\title{Check if the current R session is a Shiny application}
\usage{
isShinyApp()
}
\value{
A logical value indicating whether the current R session is a Shiny app or not.
}
\description{
This function checks if the \code{shiny} package is loaded and if the current R session is running
a Shiny application. It returns a logical value indicating whether or not the current session is
a Shiny app.
}
\examples{
isShinyApp()

}
\keyword{internal}
