% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isShinyServerApp}
\alias{isShinyServerApp}
\title{Check if the current R session is running a Shiny server application}
\usage{
isShinyServerApp()
}
\value{
\code{TRUE} if the current R session is running a Shiny server application, \code{FALSE} otherwise.
}
\description{
This function is used to determine whether the current R session is running a Shiny server application.
It internally calls the \code{\link{isHostedApp}} function to determine this.
}
\examples{
isShinyServerApp()

}
\keyword{internal}
