% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.app.R
\name{GpsSampler}
\alias{GpsSampler}
\title{GpsSampler R6 class}
\format{
A list of additional arguments
}
\value{
The launch application
}
\description{
Main entry point for the GPS sampling Shiny application.
Use the \code{\link[=sampler]{sampler()}} factory function for idiomatic construction.
}
\section{Functions}{
\itemize{
\item \code{GpsSampler}: Launch a project

}}
\section{Public}{

}

\section{Private}{

}

\section{Super classes}{
\code{\link[gpssampling:Base]{gpssampling::Base}} -> \code{gpssampling::AppShinyBase} -> \code{gpssampling::AppShinyWithAuthentification} -> \code{gpssampling::AppShinyWithTranslation} -> \code{gpssampling::AppShiny} -> \code{GpsSampler}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\code{character(1)})
A function to retrieve the application's data}

\item{\code{dlg_method}}{(\code{character(1)})
A function to retrieve the dialog method}

\item{\code{keypress}}{(\code{character(1)})
A function to retrieve the keypress}

\item{\code{data}}{(\code{character(1)})\cr}

\item{\code{dlg_method}}{(\code{character(1)})\cr}

\item{\code{keypress}}{(\code{character(1)})\cr
Initializes the object

This function initializes the object with given arguments.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GpsSampler-new}{\code{GpsSampler$new()}}
\item \href{#method-GpsSampler-launch}{\code{GpsSampler$launch()}}
\item \href{#method-GpsSampler-clone}{\code{GpsSampler$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gpssampling" data-topic="AppShinyBase" data-id="addModule"><a href='../../gpssampling/html/AppShinyBase.html#method-AppShinyBase-addModule'><code>gpssampling::AppShinyBase$addModule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gpssampling" data-topic="AppShinyBase" data-id="shutdown"><a href='../../gpssampling/html/AppShinyBase.html#method-AppShinyBase-shutdown'><code>gpssampling::AppShinyBase$shutdown()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gpssampling" data-topic="AppShinyWithAuthentification" data-id="logout"><a href='../../gpssampling/html/AppShinyWithAuthentification.html#method-AppShinyWithAuthentification-logout'><code>gpssampling::AppShinyWithAuthentification$logout()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GpsSampler-new"></a>}}
\if{latex}{\out{\hypertarget{method-GpsSampler-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GpsSampler$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A list of additional arguments

Launch a project

This function launches a project with given parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GpsSampler-launch"></a>}}
\if{latex}{\out{\hypertarget{method-GpsSampler-launch}{}}}
\subsection{Method \code{launch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GpsSampler$launch(
  open = TRUE,
  port = NULL,
  options = list(),
  test = FALSE,
  test_record = FALSE,
  authentification = FALSE,
  method = NULL,
  identify = FALSE,
  bbox = c(13.02915, 11.76533, 13.24897, 11.89713)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{open}}{logical indicating whether or not to open the project on launch}

\item{\code{port}}{Optional numeric port number for the project}

\item{\code{options}}{An optional list of named attributes that are passed as options to the project launch}

\item{\code{authentification}}{Logical indicating whether or not to use authentication}

\item{\code{method}}{An optional character string specifying a specific method for the project}

\item{\code{identify}}{Logical indicating whether or not to use identify part only}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The launch application
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GpsSampler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GpsSampler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GpsSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
