% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{launcher}
\alias{launcher}
\title{Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)}
\usage{
launcher()

launcher()
}
\description{
Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)

Create a launcher on the desktop for Windows (.bat), Mac (.command), or Linux (.sh)
}
\details{
On Windows/Mac/Linux a file named launch.bat/launch.command/launch.sh will be put on the desktop.
Double-click the file to launch the specified app app

On Windows/Mac/Linux a file named launch.bat/launch.command/launch.sh will be put on the desktop.
Double-click the file to launch the specified app app
}
\examples{
\dontrun{
launcher()
}

\dontrun{
launcher()
}

}
