% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.launcher.R
\name{lin_launcher}
\alias{lin_launcher}
\title{Create a launcher and updater for Linux (.sh)}
\usage{
lin_launcher()
}
\description{
Create a launcher and updater for Linux (.sh)
}
\details{
On Linux a file named 'launch.sh' and one named 'update.sh' will be put on the desktop. Double-click the file to launch the specified app app or update app to the latest version
}
\examples{
\dontrun{
lin_launcher()
}

}
