% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.launcher.R
\name{mac_launcher}
\alias{mac_launcher}
\title{Create a launcher and updater for Mac (.command)}
\usage{
mac_launcher()
}
\description{
Create a launcher and updater for Mac (.command)
}
\details{
On Mac a file named 'launch.command' and one named 'update.command' will be put on the desktop. Double-click the file to launch the specified app app or update app to the latest version
}
\examples{
\dontrun{
mac_launcher()
}

}
