% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.data.R
\name{matchValue}
\alias{matchValue}
\title{Function: matchValue}
\usage{
matchValue(tbl, x, var_x, var_y)
}
\arguments{
\item{tbl}{Data frame containing the data.}

\item{x}{Vector of values to match in the data frame.}

\item{var_x}{Name of the variable in the data frame to match against x.}

\item{var_y}{Name of the variable in the data frame where the matched values are retrieved from.}
}
\value{
A vector of values from var_y that match the values in x with respect to var_x.
}
\description{
Description: Function that matches values in a data frame based on given criteria.
}
\examples{
tbl <- data.frame(a = 1:10, b = rev(1:10))
matchValue(tbl, c(5L, 1L), 'a', 'b')

}
