% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{modalDialog}
\alias{modalDialog}
\title{Modal Dialog}
\usage{
modalDialog(...)
}
\arguments{
\item{...}{arguments to be passed to the \code{modalDialog} function}
}
\description{
This function creates a modal dialog using the \code{modalDialog} function from the \code{shiny} package.
}
\examples{
shiny::shinyApp(
  ui = bslib::page_fillable(
    theme = theme(),
    shiny::actionButton('show', 'Show modal dialog')
  ),
  server = function(input, output) {
    shiny::observeEvent(input$show, {
      shiny::showModal(modalDialog(
        title = 'Important message',
        icon = 'alert'
      ))
    })
  }
)

}
