% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullOrEmpty}
\alias{nullOrEmpty}
\title{Check if a value is null or empty}
\usage{
nullOrEmpty(x)
}
\arguments{
\item{x}{The value to be checked.}
}
\value{
\code{TRUE} if the value is \code{NULL} or has zero length, \code{FALSE} otherwise.
}
\description{
This function checks if a value is \code{NULL} or has zero length.
}
\examples{
nullOrEmpty(NULL)
# [1] TRUE
nullOrEmpty(character())
# [1] TRUE
nullOrEmpty(5L)
# [1] FALSE
}
\keyword{internal}
