% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{page_navbar}
\alias{page_navbar}
\title{Page Navbar}
\usage{
page_navbar(
  ...,
  title = NULL,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  bg = NULL,
  inverse = "auto",
  collapsible = TRUE,
  fluid = TRUE,
  theme = bslib::bs_theme(),
  window_title = NA,
  lang = NULL,
  h100 = getOption("epi.h100", TRUE),
  shadow = TRUE
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[bslib:page_navbar]{bslib::page_navbar}}
  \describe{
    \item{\code{title}}{A (left-aligned) title to place in the card header/footer. If
provided, other nav items are automatically right aligned.}
    \item{\code{id}}{a character string used for dynamically updating the container (see
\code{\link[bslib:nav_select]{nav_select()}}).}
    \item{\code{selected}}{a character string matching the \code{value} of a particular
\code{\link[bslib:nav_panel]{nav_panel()}} item to selected by default.}
    \item{\code{sidebar}}{A \code{\link[bslib:sidebar]{sidebar()}} component to display on every \code{\link[bslib:nav_panel]{nav_panel()}}
page.}
    \item{\code{fillable}}{Whether or not to allow \code{fill} items to grow/shrink to fit
the browser window. If \code{TRUE}, all \code{\link[bslib:nav_panel]{nav_panel()}} pages are \code{fillable}. A
character vector, matching the \code{value} of \code{\link[bslib:nav_panel]{nav_panel()}}s to be filled, may
also be provided. Note that, if a \code{sidebar} is provided, \code{fillable} makes
the main content portion fillable.}
    \item{\code{fillable_mobile}}{Whether or not \code{fillable} pages should fill the viewport's
height on mobile devices (i.e., narrow windows).}
    \item{\code{gap}}{A \link[htmltools:validateCssUnit]{CSS length unit} defining the
\code{gap} (i.e., spacing) between elements provided to \code{...}.}
    \item{\code{padding}}{Padding to use for the body. This can be a numeric vector
(which will be interpreted as pixels) or a character vector with valid CSS
lengths. The length can be between one and four. If one, then that value
will be used for all four sides. If two, then the first value will be used
for the top and bottom, while the second value will be used for left and
right. If three, then the first will be used for top, the second will be
left and right, and the third will be bottom. If four, then the values will
be interpreted as top, right, bottom, and left respectively.}
    \item{\code{position}}{Determines whether the navbar should be displayed at the top
of the page with normal scrolling behavior (\code{"static-top"}), pinned at
the top (\code{"fixed-top"}), or pinned at the bottom
(\code{"fixed-bottom"}). Note that using \code{"fixed-top"} or
\code{"fixed-bottom"} will cause the navbar to overlay your body content,
unless you add padding, e.g.: \code{tags$style(type="text/css", "body
  {padding-top: 70px;}")}}
    \item{\code{header}}{UI element(s) (\link[htmltools:builder]{htmltools::tags}) to display \emph{above} the nav
content.}
    \item{\code{footer}}{UI element(s) (\link[htmltools:builder]{htmltools::tags}) to display \emph{below} the nav
content.}
    \item{\code{bg}}{a CSS color to use for the navbar's background color.}
    \item{\code{inverse}}{Either \code{TRUE} for a light text color or \code{FALSE} for a dark
text color. If \code{"auto"} (the default), the best contrast to \code{bg} is chosen.}
    \item{\code{underline}}{Whether or not to add underline styling to page links when
active or focused.}
    \item{\code{collapsible}}{\code{TRUE} to automatically collapse the navigation
elements into an expandable menu on mobile devices or narrow window widths.}
    \item{\code{fluid}}{\code{TRUE} to use fluid layout; \code{FALSE} to use fixed
layout.}
    \item{\code{theme}}{A \code{\link[bslib:bs_theme]{bs_theme()}} object.}
    \item{\code{window_title}}{the browser window title. The default value, \code{NA}, means
to use any character strings that appear in \code{title} (if none are found, the
host URL of the page is displayed by default).}
    \item{\code{lang}}{ISO 639-1 language code for the HTML page, such as "en" or "ko".
This will be used as the lang in the \code{<html>} tag, as in \code{<html lang="en">}.
The default (NULL) results in an empty string.}
  }}
}
\value{
A \code{tag} object representing the page navbar.
}
\description{
Create a Bootstrap navbar for a page.
}
\examples{
pageNavbar(title = 'My Page', bg = 'light')

}
