% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_ignore_NA}
\alias{paste_ignore_NA}
\title{Paste Strings with Ignoring NA Values}
\usage{
paste_ignore_NA(..., sep = " ", sort = FALSE, unique = FALSE)
}
\arguments{
\item{...}{A series of character vectors that need to be concatenated.}

\item{sep}{The separator to be used between each element in the resulting string (default is a space).}

\item{sort}{Logical indicating whether to sort the elements in the resulting string (default is FALSE).}

\item{unique}{Logical indicating whether to remove duplicates from the resulting string (default is FALSE).}
}
\value{
A character vector with the concatenated strings, where any NA values are ignored.
}
\description{
This function concatenates strings together, while ignoring any NA values.
}
\examples{
paste_ignore_NA('a', 'b', NA, 'c')
# Result: 'a b c'
paste_ignore_NA('a', 'b', NA, 'c', sep = '-')
# Result: 'a-b-c'
paste_ignore_NA('a', 'b', NA, 'c', sort = TRUE)
# Result: 'a b c'
paste_ignore_NA('a', 'b', NA, 'c', unique = TRUE)
# Result: 'a b c'

}
