% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.string.R
\name{pct}
\alias{pct}
\title{Calculates the percentage of a given number}
\usage{
pct(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
A character string with the percentage of the given number
}
\description{
Calculates the percentage of a given number
}
\examples{
pct(0.5)

}
\keyword{internal}
