% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progress}
\alias{progress}
\title{Progress}
\usage{
progress(begin = FALSE, value = 1L, increment = 1L)
}
\arguments{
\item{begin}{logical. If \code{TRUE}, the progress counter is set to the initial value specified by \code{value} and \code{increment}.}

\item{value}{numeric. The initial value of the progress counter.}

\item{increment}{numeric. The amount by which the progress counter should be incremented.}
}
\description{
This function allows you to track progress by printing the current value.
By default, the function starts at 1 and increments by 1 each time it is called.
}
\examples{
progress(begin = TRUE, value = 1L, increment = 2L)
progress() # Prints '3'
progress() # Prints '5'
progress(begin = TRUE, value = 10L, increment = -1L)
progress() # Prints '9'
progress() # Prints '8'

}
\keyword{internal}
