% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progressDone}
\alias{progressDone}
\title{Progress Done}
\usage{
progressDone(..., .envir = parent.frame())
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[cli:cli_progress_bar]{cli::cli_progress_done}}
  \describe{
    \item{\code{.envir}}{The environment to use for auto-termination and for glue
substitution. It is also used to find and set the current progress bar.}
    \item{\code{id}}{Progress bar to update or terminate. If \code{NULL}, then the
current progress bar of the calling function (or \code{.envir} if
specified) is updated or terminated.}
    \item{\code{result}}{String to select successful or unsuccessful termination.
It is only used if the progress bar is not cleared from the screen.
It can be one of \code{"done"}, \code{"failed"}, \code{"clear"}, and \code{"auto"}.}
  }}
}
\description{
This function is a wrapper around the \code{cli_progress_done} function from the \code{cli} package.
It marks the progress as done and updates the progress bar accordingly.
}
\examples{
# Create a progress bar
pb <- cli::cli_progress_bar(total = 10L)
# Update the progress bar
progressDone(pb)

}
