% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progressUpdate}
\alias{progressUpdate}
\title{Updates the progress bar}
\usage{
progressUpdate(vars, text)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[cli:cli_progress_bar]{cli::cli_progress_update}}
  \describe{
    \item{\code{status}}{New status string of the progress bar, if not \code{NULL}.}
    \item{\code{total}}{Total number of progress units, or \code{NA} if it is unknown.
\code{cli_progress_update()} can update the total number of units. This is
handy if you don't know the size of a download at the beginning, and
also in some other cases. If \code{format} is set to \code{NULL}, \code{format} (plus
\code{format_done} and \code{format_failed}) will be updated when you change
\code{total} from \code{NA} to a number. I.e. default format strings will be
updated, custom ones won't be.}
    \item{\code{extra}}{Extra data to add to the progress bar. This can be
used in custom format strings for example. It should be a named list.
\code{cli_progress_update()} can update the extra data. Often you can get
away with referring to local variables in the format string, and
then you don't need to use this argument. Explicitly including these
constants or variables in \code{extra} can result in cleaner code. In
the rare cases when you need to refer to the same progress bar from
multiple functions, and you can them to \code{extra}.}
    \item{\code{.envir}}{The environment to use for auto-termination and for glue
substitution. It is also used to find and set the current progress bar.}
    \item{\code{inc}}{Increment in progress units. This is ignored if \code{set} is
not \code{NULL}.}
    \item{\code{set}}{Set the current number of progress units to this value.
Ignored if \code{NULL}.}
    \item{\code{id}}{Progress bar to update or terminate. If \code{NULL}, then the
current progress bar of the calling function (or \code{.envir} if
specified) is updated or terminated.}
    \item{\code{force}}{Whether to force a display update, even if no update is
due.}
  }}
}
\value{
None
}
\description{
This function updates the progress bar displayed by the \code{cli} package.
It passes the arguments to \code{cli_progress_update()} function from the \code{cli} package.
}
