% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{promptChoice}
\alias{promptChoice}
\title{This function prompts the user to choose from a list of options.}
\usage{
promptChoice(prompt = "Choose between: ", choice = c("Yes", "No"))
}
\arguments{
\item{prompt}{A character string representing the prompt displayed to the user. Default is 'Choose between: '.}

\item{choice}{A character vector containing the available options for the user to choose from. Default is c('Yes', 'No').}
}
\value{
The user's selected option from the provided choices as a character string, or NA if the selection is outside the valid range.
}
\description{
This function prompts the user to choose from a list of options.
}
\examples{
promptChoice() # Prompts user with default options 'Yes' and 'No'
promptChoice(prompt = 'Please select one of the following:', choice = c('Option A', 'Option B'))

}
