% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{promptIdentifier}
\alias{promptIdentifier}
\title{Prompt for an identifier}
\usage{
promptIdentifier(prompt = "Enter an identifier", default = "epi")
}
\arguments{
\item{prompt}{A character string specifying the prompt message.
Defaults to 'Enter an identifier'.}

\item{default}{A character string specifying the default value for
the identifier. Defaults to 'epi'.}
}
\value{
A valid identifier as a character string.
}
\description{
This function prompts the user to enter an identifier and validates
the input based on a specified pattern.
}
\examples{
promptIdentifier()
promptIdentifier(prompt = 'Please enter a variable name', default = 'var1')

}
\keyword{internal}
