% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{promptInteger}
\alias{promptInteger}
\title{Prompt for an Integer}
\usage{
promptInteger(prompt = "Enter an integer", range = NULL, default = 0L)
}
\arguments{
\item{prompt}{The prompt message to display to the user (default: 'Enter an integer')}

\item{default}{The default value to display in the prompt message (default: 0)}
}
\value{
The entered value as an integer
}
\description{
This function prompts the user to enter an integer value and returns the
entered value as an integer. If the entered value is not a valid integer,
the function prompts again until a valid integer is entered.
}
\examples{
promptInteger()
promptInteger('Please enter a positive integer', 1L)
}
\keyword{internal}
