% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{radians}
\alias{radians}
\title{Convert angle from degrees to radians}
\usage{
radians(angle)
}
\arguments{
\item{angle}{numeric value representing the angle in degrees}
}
\value{
numeric value representing the angle in radians
}
\description{
This function takes an angle in degrees and converts it to radians.
}
\examples{
radians(90L)

}
