% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{radioGroupButtons}
\alias{radioGroupButtons}
\title{Radio Group Buttons}
\usage{
radioGroupButtons(...)
}
\arguments{
\item{...}{Arguments passed to shinyWidgets::radioGroupButtons function.}
}
\value{
A shinyWidgets::radioGroupButtons wrapped in a shiny::div container with rounded corners.
}
\description{
This function creates a radio group input control with buttons in a Shiny app.
}
\examples{
radioGroupButtons('radio_buttons', label = 'Choose an option', choices = c('Option 1', 'Option 2', 'Option 3'))

}
