% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.pkg.R
\name{readBasemapGoogle}
\alias{readBasemapGoogle}
\title{Read Google Maps basemap tiles for given polygons}
\usage{
readBasemapGoogle(polygons, async_queue = NULL)
}
\arguments{
\item{polygons}{A spatial feature collection of polygons or multipolygons in EPSG:4326 projection}

\item{async_queue}{An optional async queue object to track progress}
}
\value{
A list of tile grids
}
\description{
This function reads Google Maps basemap tiles for the given polygons and returns a list of tile grids.
}
\examples{
# Create a polygon from North Carolina
nc <- sf::st_read(system.file('shape/nc.shp', package = 'sf'))
nc <- sf::st_transform(nc, 'EPSG:4326')

# Read Google Maps basemap tiles for the polygon
readBasemapGoogle(nc[1L, ])

}
