% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readRDSFromCache}
\alias{readRDSFromCache}
\title{Read RDS file from cache}
\usage{
readRDSFromCache(name = "config")
}
\arguments{
\item{name}{The name of the RDS file (default is 'config').}
}
\value{
The R object read from the RDS file, or NULL if the file does not exist.
}
\description{
This function reads an RDS file from the cache directory using the provided name.
If the file exists, it is read and returned as an R object. If the file does not exist,
NULL is returned.
}
\examples{
# Read RDS file from cache
data <- readRDSFromCache()
print(data)

}
\keyword{internal}
