% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace}
\alias{replace}
\title{Replace text in a file}
\usage{
replace(path, new_path, replacements)
}
\arguments{
\item{path}{A character string specifying the path to the file to be modified.}

\item{new_path}{A character string specifying the path where the modified file will be saved.}

\item{replacements}{A named character vector of patterns and their corresponding replacement values.}
}
\value{
The function writes the modified file to the specified location.
}
\description{
This function replaces specified patterns in a file with new values.
}
\examples{
# Replace 'old' with 'new' in file.txt and save as new_file.txt
replace('file.txt', 'new_file.txt', c(old = 'new'))

}
