% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.app.R
\name{sampler}
\alias{sampler}
\title{Create a GPS sampler instance}
\usage{
sampler(...)
}
\arguments{
\item{...}{Additional arguments passed to \link{GpsSampler}\verb{$new()}.}
}
\value{
A \link{GpsSampler} R6 object.
}
\description{
Factory function for creating a \link{GpsSampler} object.
This is the recommended way to instantiate the sampling application.
}
\examples{
\dontrun{
samp <- sampler()
samp$launch()
}

}
