% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveit}
\alias{saveit}
\title{Save Objects to a File}
\usage{
saveit(..., file)
}
\arguments{
\item{...}{objects to be saved}

\item{file}{character string specifying the file name or path}
}
\description{
This function saves objects to a file using the \code{save()} function in R.
}
\examples{
# save two variables to a file named 'my_data.RData'
saveit(var1, var2, file = 'my_data.RData')

}
\keyword{internal}
