% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{saveRDSToCache}
\alias{saveRDSToCache}
\title{Save RDS file to cache}
\usage{
saveRDSToCache(rds, name = "config")
}
\arguments{
\item{rds}{The R object to be saved as RDS file.}

\item{name}{The name of the RDS file in the cache directory. Default is 'config'.}
}
\description{
This function saves an RDS file to the cache directory. If the \code{rds} argument is \code{NULL}, the function will delete the RDS file from the cache if it exists.
}
\examples{
# Save an RDS object to cache
data <- iris
saveRDSToCache(data, 'mydata')

# Delete an RDS file from cache
saveRDSToCache(NULL, 'mydata')

}
\keyword{internal}
