% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.string.R
\name{score_unique}
\alias{score_unique}
\title{Calculate a score for uniqueness of elements in a character vector}
\usage{
score_unique(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A numeric score for the uniqueness of elements in \code{x}
}
\description{
This function calculates a score for the uniqueness of elements in a character vector \code{x}.
The score is a value between 0 and 1, where 0 indicates no unique elements and 1 indicates all elements are unique.
The uniqueness is determined by counting unique elements in \code{x} and the unique elements after applying the Soundex algorithm.
If \code{x} is not a character vector, the function returns 0.
}
\examples{
score_unique(c('cat', 'dog', 'cat', 'pig'))
score_unique(c('John', 'Doe', 'Jane'))
score_unique(1:10)

}
\references{
Levenshtein, V. (1966). Binary codes capable of correcting deletions, insertions, and reversals.
Soviet Physics Doklady, 10(8), 707-710.
}
\seealso{
\code{\link{stringdist::phonetic}}
}
\keyword{internal}
