% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.theming.R
\name{shinyAppMinimal}
\alias{shinyAppMinimal}
\title{shinyAppMinimal}
\usage{
shinyAppMinimal(ui, server = NULL, theme = "epicentre", port = 8000L)
}
\arguments{
\item{ui}{A shiny.tag.list object representing the UI elements of the Shiny app.}

\item{server}{A function representing the server logic of the Shiny app.}

\item{theme}{A character string specifying the theme of the app. Choices include 'bootstrap_3', 'bootstrap_5', and 'epicentre'.}

\item{port}{An integer specifying the port number for running the Shiny app.}
}
\description{
shinyAppMinimal
}
