% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.string.R
\name{snumber}
\alias{snumber}
\title{Convert values to character string with specified format}
\usage{
snumber(x, format = "\%f")
}
\arguments{
\item{x}{A vector of numeric values}

\item{format}{A character string specifying the format to use for converting
the values. Default is "\%f".}
}
\value{
A character vector with the converted values
}
\description{
This function takes a vector of numeric values and converts them to character
string using the specified format. The format can be any valid format
specification used in the \code{sprintf} function. Any occurrences of 'NA' in the
resulting string are replaced with '-'.
}
\examples{
snumber(1:5, '\%.2f')
snumber(c(1.234, NA), '\%0.3f')

}
\keyword{internal}
