% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortu}
\alias{sortu}
\title{Sort unique elements in a vector}
\usage{
sortu(x)
}
\arguments{
\item{x}{A numeric or character vector.}
}
\value{
A sorted vector containing the unique elements of \code{x}.
}
\description{
This function takes a vector as input and returns a sorted vector
containing only the unique elements of the input vector.
}
\examples{
sortu(c(3L, 1L, 2L, 2L, 3L, 1L)) # returns 1, 2, 3

}
\keyword{internal}
