% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.string.R
\name{splitMsg}
\alias{splitMsg}
\title{splitMsg}
\usage{
splitMsg(msg)
}
\arguments{
\item{msg}{The message to be split.}
}
\value{
A list containing the split components of the message.
}
\description{
This function splits a message into its components.
}
\examples{
splitMsg('INFO [2022-03-15 13:25:30] This is a sample log message.| Detail: Sample detail.|12345')

}
\keyword{internal}
