% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopf}
\alias{stopf}
\title{Stop Function}
\usage{
stopf(msg, ...)
}
\arguments{
\item{msg}{A character string representing the error message.}

\item{...}{Additional arguments to be passed to the \code{sprintf} function.}
}
\value{
This function does not return a value. It stops the execution by throwing an error.
}
\description{
This function creates an error object with the given message and stops execution.
}
\examples{
stopf('This is an error message')
stopf('Invalid input: \%s', input)

}
\keyword{internal}
