% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.string.R
\name{strReplaceAll}
\alias{strReplaceAll}
\title{Wrapper function for string replacement using \verb{[stringr::str_replace_all()]}}
\usage{
strReplaceAll(string, pattern, replacement)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described
in \link[stringi:about_search_regex]{stringi::about_search_regex}. Control options with
\code{\link[stringr:regex]{regex()}}.

For \code{str_replace_all()} this can also be a named vector
(\code{c(pattern1 = replacement1)}), in order to perform multiple replacements
in each element of \code{string}.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

You can not match boundaries, including \code{""}, with this function.}

\item{replacement}{The replacement value, usually a single string,
but it can be the a vector the same length as \code{string} or \code{pattern}.
References of the form \verb{\\1}, \verb{\\2}, etc will be replaced with
the contents of the respective matched group (created by \verb{()}).

Alternatively, supply a function (or formula): it will be passed a single
character vector and should return a character vector of the same length.

To replace the complete string with \code{NA}, use
\code{replacement = NA_character_}.}
}
\description{
Wrapper function for string replacement using \verb{[stringr::str_replace_all()]}
}
\examples{
strReplaceAll('Hello, world!', '[aeiou]', 'X')

}
