% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tctoc}
\alias{tctoc}
\title{Tictoc function}
\usage{
tctoc(expr)
}
\arguments{
\item{expr}{An expression to be evaluated.}
}
\value{
The result of evaluating the expression.
}
\description{
This function evaluates the given expression and measures the execution time.
}
\details{
This function internally uses the tic() and toc() functions from the tictoc package to measure the execution time of the given expression.
}
\keyword{internal}
