% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tctoc_msg}
\alias{tctoc_msg}
\title{Tic-Toc Message}
\usage{
tctoc_msg(tic, toc, msg)
}
\arguments{
\item{tic}{The starting time point.}

\item{toc}{The ending time point.}

\item{msg}{An optional message that will be included in the output.}
}
\value{
A character string with the formatted message including the elapsed time in milliseconds.
}
\description{
This function calculates the time elapsed between two time points (tic and toc) and returns a formatted message with the elapsed time in milliseconds.
}
\examples{
tic <- Sys.time()
# Some time-consuming operation
toc <- Sys.time()
message <- 'Operation completed successfully.'
tctoc_msg(tic, toc, message)

}
\keyword{internal}
