% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{theme}
\alias{theme}
\title{Theme function for customizing Bootstrap themes}
\usage{
theme(preview = FALSE)
}
\arguments{
\item{preview}{Logical indicating whether to preview the theme changes (default is FALSE)}
}
\value{
A customized Bootstrap theme object
}
\description{
This function allows customization of Bootstrap themes by adjusting various elements such as fonts, colors, buttons, and cards.
}
\examples{
theme()

}
