% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip}
\alias{unzip}
\title{Unzip files}
\usage{
unzip(zipfile, pattern = ".*", exdir = NULL)
}
\arguments{
\item{zipfile}{The path to the zip file to be extracted.}

\item{pattern}{A regular expression pattern to filter the extracted files.
Defaults to '.*' which matches all files.}

\item{exdir}{The path to the directory where the files should be extracted.
If NULL, a temporary directory will be created.}
}
\value{
A list with two elements:
\itemize{
\item 'files': A character vector with the paths of the extracted files.
\item 'files_count': The number of extracted files.
}
}
\description{
This function unzips a specified zip file and returns a list of extracted files and their count.
}
\examples{
# Unzip a file and get list of files
unzip('path/to/zipfile.zip')

# Unzip a file, filter by pattern, and get list of files
unzip('path/to/zipfile.zip', pattern = 'txt')

# Unzip a file and specify the extraction directory
unzip('path/to/zipfile.zip', exdir = 'path/to/exdir')

}
\keyword{internal}
