% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{useEvents}
\alias{useEvents}
\title{Add JS code Straight to your app (instead of a dependency)}
\usage{
useEvents()
}
\value{
A script
}
\description{
This function can be used in your UI to insert directly the JavaScript
functions
}
\details{
These functions are meant to be used with \code{session$sendCustomMessage}
from the server side.

\describe{
\item{showid}{Show an element with the id provided.}
\item{hideid}{Hide an element with the id provided.}
\item{showclass}{Same as showid, but with class.}
\item{hideclass}{Same as hideid, but with class.}
\item{showhref}{Same as showid, but with \verb{a[href*=}}
\item{hidehref}{Same as hideid, but with \verb{a[href*=}}
\item{clickon}{Click on an element. The full Jquery selector has to be used.}
}
}
