% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.leaflet.R
\name{useGoogle}
\alias{useGoogle}
\title{User functions :}
\usage{
useGoogle(apikey = Sys.getenv("MAPS_API_KEY_GOOGLE"))
}
\arguments{
\item{apikey}{the api-key you want to use for your application (you need to use a valid key)}
}
\value{
a tags$head that will be usefull to load the references js/css
}
\description{
Make the package more user friendly (you need to use this function into the ui part of your shinny Application)
}
