% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.app.store.R
\name{UserData}
\alias{UserData}
\title{Polygons: sf not null}
\description{
Polygons: sf not null

Polygons: sf not null
}
\section{Super classes}{
\code{\link[gpssampling:Base]{gpssampling::Base}} -> \code{gpssampling::ModShiny} -> \code{gpssampling::GpsSamplerModule} -> \code{UserData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserData-new}{\code{UserData$new()}}
\item \href{#method-UserData-updateCells}{\code{UserData$updateCells()}}
\item \href{#method-UserData-drawNewFeature}{\code{UserData$drawNewFeature()}}
\item \href{#method-UserData-getChoicesStatus}{\code{UserData$getChoicesStatus()}}
\item \href{#method-UserData-getChoicesPriority}{\code{UserData$getChoicesPriority()}}
\item \href{#method-UserData-getTileStatus}{\code{UserData$getTileStatus()}}
\item \href{#method-UserData-displayGridIdentifyStatus}{\code{UserData$displayGridIdentifyStatus()}}
\item \href{#method-UserData-displayGuidePolygon}{\code{UserData$displayGuidePolygon()}}
\item \href{#method-UserData-displayGuidePoint}{\code{UserData$displayGuidePoint()}}
\item \href{#method-UserData-export}{\code{UserData$export()}}
\item \href{#method-UserData-import}{\code{UserData$import()}}
\item \href{#method-UserData-invertSelection}{\code{UserData$invertSelection()}}
\item \href{#method-UserData-load}{\code{UserData$load()}}
\item \href{#method-UserData-cutPolygons}{\code{UserData$cutPolygons()}}
\item \href{#method-UserData-makeValidPolygons}{\code{UserData$makeValidPolygons()}}
\item \href{#method-UserData-displayPolygons}{\code{UserData$displayPolygons()}}
\item \href{#method-UserData-addPolygon}{\code{UserData$addPolygon()}}
\item \href{#method-UserData-deletePolygon}{\code{UserData$deletePolygon()}}
\item \href{#method-UserData-editPolygon}{\code{UserData$editPolygon()}}
\item \href{#method-UserData-togglePolygonSelected}{\code{UserData$togglePolygonSelected()}}
\item \href{#method-UserData-togglePolygonFocused}{\code{UserData$togglePolygonFocused()}}
\item \href{#method-UserData-setStylePolygonFocused}{\code{UserData$setStylePolygonFocused()}}
\item \href{#method-UserData-addProject}{\code{UserData$addProject()}}
\item \href{#method-UserData-projectClone}{\code{UserData$projectClone()}}
\item \href{#method-UserData-projectEdit}{\code{UserData$projectEdit()}}
\item \href{#method-UserData-projectDelete}{\code{UserData$projectDelete()}}
\item \href{#method-UserData-projectSelect}{\code{UserData$projectSelect()}}
\item \href{#method-UserData-addRoof}{\code{UserData$addRoof()}}
\item \href{#method-UserData-removeRoof}{\code{UserData$removeRoof()}}
\item \href{#method-UserData-roofsAddOSM}{\code{UserData$roofsAddOSM()}}
\item \href{#method-UserData-addRoofs}{\code{UserData$addRoofs()}}
\item \href{#method-UserData-addSample}{\code{UserData$addSample()}}
\item \href{#method-UserData-addSamples}{\code{UserData$addSamples()}}
\item \href{#method-UserData-displayRoofs}{\code{UserData$displayRoofs()}}
\item \href{#method-UserData-generateSamples}{\code{UserData$generateSamples()}}
\item \href{#method-UserData-save}{\code{UserData$save()}}
\item \href{#method-UserData-clear}{\code{UserData$clear()}}
\item \href{#method-UserData-rollback}{\code{UserData$rollback()}}
\item \href{#method-UserData-readFromCache}{\code{UserData$readFromCache()}}
\item \href{#method-UserData-saveToCache}{\code{UserData$saveToCache()}}
\item \href{#method-UserData-sampleAppend}{\code{UserData$sampleAppend()}}
\item \href{#method-UserData-calculateSample}{\code{UserData$calculateSample()}}
\item \href{#method-UserData-displaySample}{\code{UserData$displaySample()}}
\item \href{#method-UserData-sampleDisplayToTable}{\code{UserData$sampleDisplayToTable()}}
\item \href{#method-UserData-sampleExportFilename}{\code{UserData$sampleExportFilename()}}
\item \href{#method-UserData-sampleExportToDoc}{\code{UserData$sampleExportToDoc()}}
\item \href{#method-UserData-sampleExportToDocEach}{\code{UserData$sampleExportToDocEach()}}
\item \href{#method-UserData-sampleQuadratEdit}{\code{UserData$sampleQuadratEdit()}}
\item \href{#method-UserData-sampleQuadratPost}{\code{UserData$sampleQuadratPost()}}
\item \href{#method-UserData-sampleLoadTable}{\code{UserData$sampleLoadTable()}}
\item \href{#method-UserData-sampleToggleSelect}{\code{UserData$sampleToggleSelect()}}
\item \href{#method-UserData-samplesSetValueFromTable}{\code{UserData$samplesSetValueFromTable()}}
\item \href{#method-UserData-samplesSetValueFromMap}{\code{UserData$samplesSetValueFromMap()}}
\item \href{#method-UserData-addCell}{\code{UserData$addCell()}}
\item \href{#method-UserData-searchSwipeTile}{\code{UserData$searchSwipeTile()}}
\item \href{#method-UserData-validAll}{\code{UserData$validAll()}}
\item \href{#method-UserData-triggerProjectName}{\code{UserData$triggerProjectName()}}
\item \href{#method-UserData-invalidateGrid}{\code{UserData$invalidateGrid()}}
\item \href{#method-UserData-clone}{\code{UserData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="gpssampling" data-topic="ModShiny" data-id="bind"><a href='../../gpssampling/html/ModShiny.html#method-ModShiny-bind'><code>gpssampling::ModShiny$bind()</code></a></span></li>
<li><span class="pkg-link" data-pkg="gpssampling" data-topic="ModShiny" data-id="getUI"><a href='../../gpssampling/html/ModShiny.html#method-ModShiny-getUI'><code>gpssampling::ModShiny$getUI()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-new}{}}}
\subsection{Method \code{new()}}{
Initialize Application class for shiny app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$new(application, load = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new Application object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-updateCells"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-updateCells}{}}}
\subsection{Method \code{updateCells()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$updateCells(polygons_sf)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-drawNewFeature"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-drawNewFeature}{}}}
\subsection{Method \code{drawNewFeature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$drawNewFeature(map, feature)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-getChoicesStatus"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-getChoicesStatus}{}}}
\subsection{Method \code{getChoicesStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$getChoicesStatus()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-getChoicesPriority"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-getChoicesPriority}{}}}
\subsection{Method \code{getChoicesPriority()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$getChoicesPriority()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-getTileStatus"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-getTileStatus}{}}}
\subsection{Method \code{getTileStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$getTileStatus(polygon = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-displayGridIdentifyStatus"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-displayGridIdentifyStatus}{}}}
\subsection{Method \code{displayGridIdentifyStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$displayGridIdentifyStatus(map, token, force = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-displayGuidePolygon"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-displayGuidePolygon}{}}}
\subsection{Method \code{displayGuidePolygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$displayGuidePolygon(map)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-displayGuidePoint"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-displayGuidePoint}{}}}
\subsection{Method \code{displayGuidePoint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$displayGuidePoint(map)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-export"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-export}{}}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$export()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-import"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-import}{}}}
\subsection{Method \code{import()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-invertSelection"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-invertSelection}{}}}
\subsection{Method \code{invertSelection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$invertSelection(map)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-load"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-load}{}}}
\subsection{Method \code{load()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$load(project_active = NULL, force = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-cutPolygons"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-cutPolygons}{}}}
\subsection{Method \code{cutPolygons()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$cutPolygons(map, line)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-makeValidPolygons"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-makeValidPolygons}{}}}
\subsection{Method \code{makeValidPolygons()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$makeValidPolygons(sf, id_max = 0L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-displayPolygons"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-displayPolygons}{}}}
\subsection{Method \code{displayPolygons()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$displayPolygons(
  map,
  only_selected = FALSE,
  sample = FALSE,
  cells = FALSE,
  plan = FALSE,
  fit = TRUE,
  status = TRUE,
  sf = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-addPolygon"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-addPolygon}{}}}
\subsection{Method \code{addPolygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$addPolygon(map, polygon)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-deletePolygon"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-deletePolygon}{}}}
\subsection{Method \code{deletePolygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$deletePolygon(map, id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-editPolygon"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-editPolygon}{}}}
\subsection{Method \code{editPolygon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$editPolygon(map, feature)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-togglePolygonSelected"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-togglePolygonSelected}{}}}
\subsection{Method \code{togglePolygonSelected()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$togglePolygonSelected(map, id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-togglePolygonFocused"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-togglePolygonFocused}{}}}
\subsection{Method \code{togglePolygonFocused()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$togglePolygonFocused(map, id, group, step)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-setStylePolygonFocused"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-setStylePolygonFocused}{}}}
\subsection{Method \code{setStylePolygonFocused()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$setStylePolygonFocused(map, focused, group = "polygons")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-addProject"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-addProject}{}}}
\subsection{Method \code{addProject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$addProject(status, priority, name, description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-projectClone"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-projectClone}{}}}
\subsection{Method \code{projectClone()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$projectClone(status, priority, name, description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-projectEdit"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-projectEdit}{}}}
\subsection{Method \code{projectEdit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$projectEdit(status, priority, description)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-projectDelete"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-projectDelete}{}}}
\subsection{Method \code{projectDelete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$projectDelete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-projectSelect"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-projectSelect}{}}}
\subsection{Method \code{projectSelect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$projectSelect(name = "default")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-addRoof"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-addRoof}{}}}
\subsection{Method \code{addRoof()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$addRoof(map, lng, lat, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-removeRoof"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-removeRoof}{}}}
\subsection{Method \code{removeRoof()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$removeRoof(map, roof_id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-roofsAddOSM"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-roofsAddOSM}{}}}
\subsection{Method \code{roofsAddOSM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$roofsAddOSM(map, progress, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-addRoofs"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-addRoofs}{}}}
\subsection{Method \code{addRoofs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$addRoofs(map, roofs, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-addSample"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-addSample}{}}}
\subsection{Method \code{addSample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$addSample(map = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-addSamples"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-addSamples}{}}}
\subsection{Method \code{addSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$addSamples(map, sf, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-displayRoofs"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-displayRoofs}{}}}
\subsection{Method \code{displayRoofs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$displayRoofs(map, roofs, tile_x = roofs$x[1L], tile_y = roofs$y[1L])}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-generateSamples"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-generateSamples}{}}}
\subsection{Method \code{generateSamples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$generateSamples(
  map = NULL,
  type = "SP_QDR",
  method = "random",
  count = 35L,
  size = 25L,
  polygon_idx = NULL,
  samples = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-save"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$save(
  delimit = FALSE,
  identify = FALSE,
  sample = FALSE,
  result = FALSE,
  project = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-clear"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$clear(
  delimit = FALSE,
  identify = FALSE,
  sample = FALSE,
  result = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-rollback"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-rollback}{}}}
\subsection{Method \code{rollback()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$rollback(
  delimit = FALSE,
  identify = FALSE,
  sample = FALSE,
  result = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-readFromCache"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-readFromCache}{}}}
\subsection{Method \code{readFromCache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$readFromCache(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-saveToCache"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-saveToCache}{}}}
\subsection{Method \code{saveToCache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$saveToCache(rds, file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleAppend"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleAppend}{}}}
\subsection{Method \code{sampleAppend()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleAppend(map = NULL, polygon_idx = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-calculateSample"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-calculateSample}{}}}
\subsection{Method \code{calculateSample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$calculateSample(progress = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-displaySample"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-displaySample}{}}}
\subsection{Method \code{displaySample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$displaySample(
  map,
  polygon_idx,
  samples_idx = NULL,
  plan = FALSE,
  fit = FALSE,
  tooltip = FALSE,
  status = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleDisplayToTable"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleDisplayToTable}{}}}
\subsection{Method \code{sampleDisplayToTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleDisplayToTable(
  table_id,
  polygon_idx,
  samples_idx = NULL,
  plan = FALSE,
  tooltip = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleExportFilename"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleExportFilename}{}}}
\subsection{Method \code{sampleExportFilename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleExportFilename(extension)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleExportToDoc"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleExportToDoc}{}}}
\subsection{Method \code{sampleExportToDoc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleExportToDoc(file, open = TRUE, polygon_idx = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleExportToDocEach"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleExportToDocEach}{}}}
\subsection{Method \code{sampleExportToDocEach()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleExportToDocEach(
  file,
  open = TRUE,
  polygon_idx = NULL,
  samples_idxs = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleQuadratEdit"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleQuadratEdit}{}}}
\subsection{Method \code{sampleQuadratEdit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleQuadratEdit(map, id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleQuadratPost"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleQuadratPost}{}}}
\subsection{Method \code{sampleQuadratPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleQuadratPost(map, feature)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleLoadTable"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleLoadTable}{}}}
\subsection{Method \code{sampleLoadTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleLoadTable(tbl_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-sampleToggleSelect"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-sampleToggleSelect}{}}}
\subsection{Method \code{sampleToggleSelect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$sampleToggleSelect(map, sample)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-samplesSetValueFromTable"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-samplesSetValueFromTable}{}}}
\subsection{Method \code{samplesSetValueFromTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$samplesSetValueFromTable(map, recs, vars, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-samplesSetValueFromMap"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-samplesSetValueFromMap}{}}}
\subsection{Method \code{samplesSetValueFromMap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$samplesSetValueFromMap(table_id, id, var, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-addCell"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-addCell}{}}}
\subsection{Method \code{addCell()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$addCell(map, token, cell = NULL, inverse = FALSE, toggle = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-searchSwipeTile"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-searchSwipeTile}{}}}
\subsection{Method \code{searchSwipeTile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$searchSwipeTile(
  map,
  direction,
  map_size,
  token,
  mapped = FALSE,
  valid = FALSE,
  invalid = FALSE,
  lat = NULL,
  lon = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-validAll"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-validAll}{}}}
\subsection{Method \code{validAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$validAll(map)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-triggerProjectName"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-triggerProjectName}{}}}
\subsection{Method \code{triggerProjectName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$triggerProjectName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-invalidateGrid"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-invalidateGrid}{}}}
\subsection{Method \code{invalidateGrid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$invalidateGrid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
