% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{useWaiter}
\alias{useWaiter}
\title{Use Waiter and Waitress in Shiny app}
\usage{
useWaiter(particle = FALSE)
}
\value{
A tag list containing Waiter and Waitress components for a Shiny app.
}
\description{
This function integrates Waiter and Waitress into a Shiny application.
}
\examples{
useWaiter()

}
