% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\title{Functions usefull}
\description{
MyPackage has different families of functions described below
}
\section{Some functions}{

\itemize{
\item \code{\link[=getPackageDescription]{getPackageDescription()}}
\item \code{\link[=getPackageField]{getPackageField()}}
}
}

\section{Other functions}{

\itemize{
\item \code{\link[=getPackageField]{getPackageField()}}
\item \code{\link[=getPackageField]{getPackageField()}}
}
}

