% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.paths.R
\name{utils_dir}
\alias{utils_dir}
\alias{getDirApp}
\alias{getDirAppAll}
\alias{getDirAppAllData}
\alias{getDirAppSharepoint}
\alias{getDirAppTemp}
\alias{getDirAppUsers}
\alias{getDirApps}
\alias{getDirAssets}
\alias{getDirDatabase}
\alias{getDirDevelopment}
\alias{getDirOrganization}
\alias{getDirPackage}
\alias{getDirSharepoint}
\alias{getDirShiny}
\alias{getDirTemp}
\alias{getDirUser}
\alias{getDirUserData}
\alias{getDirUserDataLocal}
\alias{getDirUserDataRoaming}
\alias{getDirUserHome}
\alias{createDir}
\title{Functions Package Dir}
\usage{
getDirApp(...)

getDirAppAll(...)

getDirAppAllData(...)

getDirAppSharepoint(path = "MSF", ...)

getDirAppTemp(...)

getDirAppUsers(...)

getDirApps(...)

getDirAssets(...)

getDirDatabase()

getDirDevelopment(...)

getDirOrganization(org = "epicentre", ...)

getDirPackage(...)

getDirSharepoint(path = "MSF")

getDirShiny(...)

getDirTemp(...)

getDirUser(...)

getDirUserData(...)

getDirUserDataLocal(...)

getDirUserDataRoaming(...)

getDirUserHome()

createDir(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[fs:path]{fs::path}}
  \describe{
    \item{\code{ext}}{An optional extension to append to the generated path.}
  }}

\item{path}{Character string indicating the directory inside the SharePoint folder (default is 'MSF')}

\item{org}{The name of the organization for which the directory is to be created. Default is 'epicentre'.}
}
\value{
A character string representing the full path to the user's roaming data directory.
}
\description{
This function returns the directory path where all the data files of the application are stored.
}
\section{Functions}{
\itemize{
\item \code{getDirApp()}: Get Directory Data

This function returns the path to the user data directory for the package.

\item \code{getDirAppAll()}: Get the directory path for all application directories.

This function creates a directory with the name 'all' inside the main application directory
and returns the full path to this new directory.

\item \code{getDirAppAllData()}: Get the directory path for all data files in the application.

\item \code{getDirAppSharepoint()}: Get the directory path in the SharePoint application

This function returns the directory path in the SharePoint application folder.

\item \code{getDirAppTemp()}: Get temporary directory under the app directory

This function creates a temporary directory under the application directory and returns the path to that directory.

\item \code{getDirAppUsers()}: Get the directory path for application users

This function returns the absolute path of the directory for application users.
The function utilizes the \code{\link{createDir}} function to create the users directory if it doesn't exist.

\item \code{getDirApps()}: Get directory for storing applications

This function creates a directory for storing applications within the organization directory.

\item \code{getDirAssets()}: Chemin des ressource du package (image, icon, fonts, css, js, ...)

Cette fonction renvoie le chemin du repertoire contenant les ressources du package,
telles que les images, les icones, les polices, les fichiers CSS, les fichiers JS, etc.

\item \code{getDirDatabase()}: Get the directory path of the database

This function retrieves the directory path where the database is located.

\item \code{getDirDevelopment()}: Get the path to the development directory.

This function retrieves the path to the development directory, based on the operating system.
For Windows, it first obtains the user library directory using the \code{Sys.getenv} function,
and then applies the \code{fs::path_dir} function multiple times to navigate to the development directory.
For other operating systems, it also retrieves the user library directory using \verb{Sys.getenv', and applies the }fs::path_dir` function to navigate to the development directory.
Finally, it constructs the full path by appending any additional subdirectories specified in the arguments.

\item \code{getDirOrganization()}: Get Directory Organization

This function creates a directory with the specified organization name and returns its path.

\item \code{getDirPackage()}: Get directory path within the package

This function returns the path of a directory within the package.

\item \code{getDirSharepoint()}: Get the SharePoint directory path

This function returns the path of the SharePoint directory.

\item \code{getDirShiny()}: Get Shiny directory

This function returns the directory where the Shiny application is hosted.

\item \code{getDirTemp()}: Get Temporary Directory Path

This function returns the absolute path to the temporary directory.

\item \code{getDirUser()}: Directory of the current user

This function returns the directory of the current user.

\item \code{getDirUserData()}: Get the directory for user data

This function returns the directory path for user data. It first checks the operating system. If it is Windows, it calls the function \code{getDirUserDataLocal()} from the \code{fs} package to get the local user data directory. Otherwise, it returns \code{NULL}.

\item \code{getDirUserDataLocal()}: Get local user data directory

This function returns the path to the local user data directory. It uses the
\code{rappdirs::user_data_dir} function to get the user data directory and appends
any additional path provided.

\item \code{getDirUserDataRoaming()}: Get the path to the user's roaming data directory.

\item \code{getDirUserHome()}: Get user's home directory

This function returns the path of the user's home directory.

\item \code{createDir()}: Create Directory

This function creates a directory at the specified path. If the directory already exists, it does nothing.

}}
\examples{
getDirData()
getDirData('subdir')

getDirApps()
getDirApps('app1')

getDirAppAllData()

getDirAppSharepoint()

getDirAppTemp()
getDirAppTemp('subdir')

getDirApps() # creates a directory called 'apps' within the organization directory
getDirApps('app1') # creates a directory called 'app1' within the organization directory

getDirOrganization()
getDirOrganization(org = 'myOrg')

getDirPackage('data')
getDirPackage('inst', 'templates')

getDirSharepoint()
# Returns: 'C:/Users/<Username>/OneDrive/Sharepoint/MSF'

getDirSharepoint('Documents')
# Returns: 'C:/Users/<Username>/OneDrive/Sharepoint/Documents'

getDirShiny()

getDirUser()

getDirUserDataLocal()
getDirUserDataLocal('myapp')

getDirUserDataRoaming('subdir1', 'subdir2')
# Returns 'C:/Users/User/AppData/Roaming/subdir1/subdir2'

getDirUserDataRoaming('filename.txt')
# Returns 'C:/Users/User/AppData/Roaming/filename.txt'

createDir('path/to/directory')

}
\seealso{
\code{\link{getDirApp}}, \code{\link{createDir}}

\code{\link{createDir}}, \code{\link{getDirApp}}
}
\keyword{internal}
