% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_package}
\alias{utils_package}
\alias{getPackageDescription}
\title{Functions Package}
\usage{
getPackageDescription(fields = NULL)
}
\arguments{
\item{fields}{A character vector specifying the fields to extract from the package description.
If not specified, all fields will be returned.}
}
\value{
A character vector containing the requested fields from the package description.
}
\description{
MyPackage has different families of functions described below

Return the description of the current package.
}
\details{
This function uses the \code{utils::packageName()} function to get the name of the current package,
and then calls \code{utils::packageDescription()} to retrieve the package description. The \code{fields} parameter
can be used to specify which fields to extract from the package description. If \code{fields} is not specified,
all fields will be returned.
}
\examples{
\dontrun{
desc <- getPackageDescription()
print(desc)
}

}
\keyword{internal}
