% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.launcher.R
\name{win_launcher}
\alias{win_launcher}
\title{Create a launcher and updater for Windows (.bat)}
\usage{
win_launcher()
}
\description{
Create a launcher and updater for Windows (.bat)
}
\details{
On Windows a file named 'launch.bat' and one named 'update.bat' will be put on the desktop. Double-click the file to launch the specified app app or update app to the latest version
}
\examples{
\dontrun{
win_launcher()
}

}
