% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.string.R
\name{writeUtf8}
\alias{writeUtf8}
\title{Write character vector as a UTF-8 encoded file}
\usage{
writeUtf8(x, file, bom = FALSE)
}
\arguments{
\item{x}{A character vector to be written to the file.}

\item{file}{The path to the file to be written.}

\item{bom}{A logical value indicating whether a Byte Order Mark (BOM) should be included in the file. Default is \code{FALSE}.}
}
\description{
This function writes a character vector as a UTF-8 encoded file.
}
\examples{
# Create a character vector
text <- c('Hello', 'world!')

# Write the character vector to a file without BOM
writeUtf8(text, 'output.txt')

# Write the character vector to a file with BOM
writeUtf8(text, 'output.txt', bom = TRUE)

}
\keyword{internal}
