% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.shiny.R
\name{findFreePort}
\alias{findFreePort}
\title{Find a free port to use}
\usage{
findFreePort(start = 1000L, random = FALSE)
}
\arguments{
\item{start}{The starting port number to begin searching from}

\item{random}{Logical indicating whether to return a random available port or the first available port}
}
\value{
The first available free port number
}
\description{
Find a free port to use
}
\examples{
findFreePort(6000L, TRUE)
findFreePort(6000L)

}
