% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getShinyServerHost}
\alias{getShinyServerHost}
\title{Get the Shiny Server host}
\usage{
getShinyServerHost(session)
}
\arguments{
\item{session}{The Shiny session object.}
}
\value{
The Shiny Server host.
}
\description{
Retrieves the Shiny Server host based on the session object.
}
\details{
If the app is hosted, the Shiny Server host is set to 'https://apps.msf.net'.
Otherwise, the Shiny Server host is set to 'https://apps.msf.net'.
}
\keyword{internal}
