% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.app.R
\name{Application}
\alias{Application}
\title{Application R6 class}
\format{
A list of additional arguments
}
\value{
The launch application
}
\description{
This class represents an application with various methods and functionalities.
}
\section{Functions}{
\itemize{
\item \code{Application}: Launch a project

}}
\section{Public}{

}

\section{Private}{

}

\section{Super classes}{
\code{epi.geosampler::Base} -> \code{epi.geosampler::AppShinyBase} -> \code{epi.geosampler::AppShinyWithAuthentification} -> \code{epi.geosampler::AppShinyWithTranslation} -> \code{epi.geosampler::AppShiny} -> \code{Application}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\code{character(1)})
A function to retrieve the application's data}

\item{\code{dlg_method}}{(\code{character(1)})
A function to retrieve the dialog method}

\item{\code{keypress}}{(\code{character(1)})
A function to retrieve the keypress}

\item{\code{data}}{(\code{character(1)})\cr}

\item{\code{dlg_method}}{(\code{character(1)})\cr}

\item{\code{keypress}}{(\code{character(1)})\cr
Initializes the object

This function initializes the object with given arguments.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Application-new}{\code{Application$new()}}
\item \href{#method-Application-launch}{\code{Application$launch()}}
\item \href{#method-Application-clone}{\code{Application$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="epi.geosampler" data-topic="AppShinyBase" data-id="addModule"><a href='../../epi.geosampler/html/AppShinyBase.html#method-AppShinyBase-addModule'><code>epi.geosampler::AppShinyBase$addModule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="epi.geosampler" data-topic="AppShinyBase" data-id="shutdown"><a href='../../epi.geosampler/html/AppShinyBase.html#method-AppShinyBase-shutdown'><code>epi.geosampler::AppShinyBase$shutdown()</code></a></span></li>
<li><span class="pkg-link" data-pkg="epi.geosampler" data-topic="AppShinyWithAuthentification" data-id="logout"><a href='../../epi.geosampler/html/AppShinyWithAuthentification.html#method-AppShinyWithAuthentification-logout'><code>epi.geosampler::AppShinyWithAuthentification$logout()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Application-new"></a>}}
\if{latex}{\out{\hypertarget{method-Application-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{A list of additional arguments

Launch a project

This function launches a project with given parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Application-launch"></a>}}
\if{latex}{\out{\hypertarget{method-Application-launch}{}}}
\subsection{Method \code{launch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$launch(
  open = TRUE,
  port = NULL,
  options = list(),
  test = FALSE,
  test_record = FALSE,
  authentification = FALSE,
  method = NULL,
  identify = FALSE,
  bbox = c(13.02915, 11.76533, 13.24897, 11.89713)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{open}}{logical indicating whether or not to open the project on launch}

\item{\code{port}}{Optional numeric port number for the project}

\item{\code{options}}{An optional list of named attributes that are passed as options to the project launch}

\item{\code{authentification}}{Logical indicating whether or not to use authentication}

\item{\code{method}}{An optional character string specifying a specific method for the project}

\item{\code{identify}}{Logical indicating whether or not to use identify part only}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The launch application
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Application-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Application-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
